Fix TX conflicting, transaction not in the memory pool
When add pending wallet transactions, it will try to add the other pending wallet tx back to the memory pool, as well as when it's added out of order, it appears as if its orphans with missing inputs. Those transactions are then rejected and flagged as conflicting (0/unconfirmed, not in memory pool). To avoid this, we need to arrange those transactions in the specific manner