When the new block is being accepted, we check all transactions. If the transaction is coinbase or coinstake, then we attempt to find the reward for the masternode in it.
For PoS blocks - 40% bonus before hardening and 20% after.

For PoW blocks - 20% bonus after hardfork.

There is also the concept that the accepted block may not be checked (when re-indexing) since we don't know which masternodes were active during the period of the block is accepted. So we need a list of active masternode to find the reward in the transaction.

Therefore, this code should not reject accepted blocks even when re-indexed even if they don't have rewards for the masternode.

Signed-off-by: 216k155 <216k155@luxcore.io>