Fix: Invalid memory access on shutdown
Problem:
When the wallet is rescanning block for its transaction, if we close
the wallet, the wallet will be crashed and wallet.dat will be corrupted
in some case.

The root cause is that the rescanning process is still on progress while
the wallet is shutting down. When OS releases all allocated memory of the
wallet and the scanning process accesses that memory, invalid memory access
will occur => wallet.dat will be corrupted in this case

Signed-off-by: tran <nguyenhoangtran11@yahoo.com>

(cherry picked from commit bac66b0)

(cherry picked from commit afbd5be)