Coins - Fix Assertion it->second.flags &CCoinsCacheEntry::FRESH failed by Alter assumptions in CCoinsViewCache::BatchWrite
 The spendcoin consume the output. So whether the incoming moveout pointer is empty, whether to save the corresponding coin object. If the found entry status is fresh, it gonna show u that it is consistent with the backend view, delete the entry, otherwise the entry is dirty. Set the it value to -1
The coinsviewcache batch..... method is which mainly serves the coinviewcache object as the backend view of another coinview object, accepts the state of the upper view through the coinsmap object, and updates the internal state of the object. skip non-DIRTY entries when iterating the upper view, forexample the entry that has not been modified. if an entry is found in the upper view, not found inside the object, skip the FRESH and consumed item, and the rest is refreshed to this view; if found, the entry is The upper view is fresh and consumed. In this view, the entry with the key is deleted. Otherwise, the coin in the entry is updated and the entry is marked DIRTY
Signed-off-by: 216k155 <216k155@luxcore.io>