txdb: increase default total db memory cache (-dbcache)
bitcoin default is now 300 MB, but we don't need so much

we need a bit more memory for the 45% of these 256MB used for the coins cache,
especially if indexes are used. Current 45MB were too small for the normal initial db check.

This change can also increase the overall stability.